refactored mealplan interface adapters and view
