+++ src/interface_adapters/mealplan/MealPlanPresenter.java
@@ -1,4 +1,4 @@
-package src.interfaceadaptersmealplan;
+package src.interface_adapters.mealplan;
 
 import src.use_case.mealplan.MealPlanOutputBoundary;
 import src.use_case.mealplan.MealPlanOutputData;
